% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qs.R
\name{qs-terra}
\alias{qs-terra}
\alias{qreadvect}
\alias{qreadext}
\alias{qsavevect}
\alias{qsaveext}
\title{\code{qread} and \code{qsave} routines for \code{terra}}
\usage{
qreadvect(...)

qreadext(...)

qsavevect(x, file, ...)

qsaveext(x, file, ...)
}
\arguments{
\item{x, file, ...}{Arguments.}
}
\description{
These functions provide \code{\link[qs:qread]{qs::qread()}} and \code{\link[qs:qsave]{qs::qsave()}} wrappers for \code{terra} objects.
}
