% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatRaster.R
\name{spat}
\alias{spat}
\alias{spatNormalise}
\alias{overwriteRaster}
\title{\code{SpatRaster} routines}
\usage{
spatNormalise(x)

overwriteRaster(x, filename, ...)
}
\arguments{
\item{x, filename, ...}{Arguments.}
}
\description{
These functions provide \code{SpatRaster} routines.
}
