% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.extensions-package.R
\docType{package}
\name{spatial.extensions}
\alias{spatial.extensions-package}
\alias{spatial.extensions}
\title{\code{\link{spatial.extensions}}}
\description{
Extensions to r-spatial packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwardlavender/spatial.extensions/}
  \item Report bugs at \url{https://github.com/edwardlavender/spatial.extensionss/issues/}
}

}
\author{
\strong{Maintainer}: Edward Lavender \email{edward.lavender@eawag.ch} (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})

}
